+++ styx-scheduler-service/src/test/java/com/spotify/styx/SchedulerTest.java
@@ -24,23 +24,11 @@ import static com.spotify.styx.state.TimeoutConfig.createWithDefaultTtl;
 import static java.time.Duration.ofSeconds;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.sameInstance;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyBoolean;
-import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.only;
-import static org.mockito.Mockito.timeout;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
-import com.google.api.client.repackaged.com.google.common.base.Throwables;
 import com.google.common.collect.Lists;
-import com.spotify.styx.model.Backfill;
 import com.spotify.styx.model.Event;
 import com.spotify.styx.model.Resource;
 import com.spotify.styx.model.Schedule;
@@ -54,23 +42,16 @@ import com.spotify.styx.state.StateData;
 import com.spotify.styx.state.StateManager;
 import com.spotify.styx.state.SyncStateManager;
 import com.spotify.styx.state.TimeoutConfig;
-import com.spotify.styx.state.Trigger;
 import com.spotify.styx.storage.Storage;
-import com.spotify.styx.util.ParameterUtil;
 import com.spotify.styx.util.Time;
 import java.io.IOException;
 import java.net.URI;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.TimeUnit;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -86,38 +67,17 @@ public class SchedulerTest {
   private static final WorkflowInstance INSTANCE =
       WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T01");
 
-  private static final Backfill BACKFILL_1 = Backfill.newBuilder()
-      .id("backfill-1")
-      .start(Instant.parse("2016-12-02T22:00:00Z"))
-      .end(Instant.parse("2016-12-05T22:00:00Z"))
-      .workflowId(WORKFLOW_ID1)
-      .concurrency(2)
-      .schedule(Schedule.HOURS)
-      .nextTrigger(Instant.parse("2016-12-02T22:00:00Z"))
-      .build();
+  private WorkflowCache workflowCache;
+  private Storage storage;
+  private StateManager stateManager;
+  private Scheduler scheduler;
 
-  private static final Backfill BACKFILL_2 = Backfill.newBuilder()
-      .id("backfill-2")
-      .start(Instant.parse("2016-12-02T00:00:00Z"))
-      .end(Instant.parse("2016-12-02T03:00:00Z"))
-      .workflowId(WORKFLOW_ID1)
-      .concurrency(3)
-      .schedule(Schedule.HOURS)
-      .nextTrigger(Instant.parse("2016-12-02T00:00:00Z"))
-      .build();
+  private Instant now = Instant.parse("2016-12-02T22:00:00Z");
+  private Time time = () -> now;
 
-  WorkflowCache workflowCache;
-  Storage storage;
-  StateManager stateManager;
-  Scheduler scheduler;
-  TriggerListener triggerListener;
+  private List<Resource> resourceLimits = Lists.newArrayList();
 
-  Instant now = Instant.parse("2016-12-02T22:00:00Z");
-  Time time = () -> now;
-
-  List<Resource> resourceLimits = Lists.newArrayList();
-
-  ExecutorService executor = Executors.newCachedThreadPool();
+  private ExecutorService executor = Executors.newCachedThreadPool();
 
   @After
   public void tearDown() throws Exception {
@@ -129,16 +89,11 @@ public class SchedulerTest {
     TimeoutConfig timeoutConfig = createWithDefaultTtl(ofSeconds(timeoutSeconds));
 
     storage = mock(Storage.class);
-    triggerListener = mock(TriggerListener.class);
     when(storage.resources()).thenReturn(resourceLimits);
     when(storage.globalConcurrency()).thenReturn(Optional.empty());
 
     stateManager = new SyncStateManager();
-    scheduler = new Scheduler(time, timeoutConfig, stateManager, workflowCache, storage,
-                              triggerListener);
-
-    when(triggerListener.event(any(Workflow.class), any(Trigger.class), any(Instant.class)))
-        .then(a -> CompletableFuture.completedFuture(null));
+    scheduler = new Scheduler(time, timeoutConfig, stateManager, workflowCache, storage);
   }
 
   private void setResourceLimit(String resourceId, long limit) {
@@ -154,7 +109,7 @@ public class SchedulerTest {
     stateManager.initialize(runState);
   }
 
-  Workflow workflowUsingResources(WorkflowId id, String... resources) {
+  private Workflow workflowUsingResources(WorkflowId id, String... resources) {
     return Workflow.create(
         id.componentId(),
         URI.create("http://example.com"),
@@ -163,194 +118,6 @@ public class SchedulerTest {
             .schedule(Schedule.HOURS)
             .resources(resources)
             .build());
-  }
-
-  @Test
-  public void shouldTriggerBackfillsNew() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    final int concurrency = BACKFILL_1.concurrency();
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(BACKFILL_1));
-
-    scheduler.tick();
-
-    final List<Instant> instants =
-        ParameterUtil.rangeOfInstants(BACKFILL_1.start(), BACKFILL_1.end(),
-                                      workflow.configuration().schedule());
-
-    instants.stream().limit(concurrency).forEach(
-        instant ->
-            verify(triggerListener).event(workflow, Trigger.backfill(BACKFILL_1.id()), instant));
-
-    verify(storage)
-        .storeBackfill(BACKFILL_1.builder().nextTrigger(instants.get(concurrency)).build());
-  }
-
-  @Test
-  public void shouldTriggerBackfillsInProgress() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    when(storage.backfills(anyBoolean())).thenReturn(
-        Collections.singletonList(BACKFILL_1.builder()
-            .nextTrigger(Instant.parse("2016-12-03T00:00:00Z"))
-            .build()));
-
-    stateManager.initialize
-        (RunState.fresh(
-            WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T23")));
-    stateManager.receive(
-        Event.triggerExecution(
-            WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T23"),
-            Trigger.backfill("backfill-1")));
-
-    scheduler.tick();
-
-    verify(triggerListener, only()).event(
-        workflow,
-        Trigger.backfill(BACKFILL_1.id()), Instant.parse("2016-12-03T00:00:00Z"));
-  }
-
-  @Test
-  public void shouldTriggerBackfillsToCompletion() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(BACKFILL_2));
-
-    scheduler.tick();
-
-    List<Instant> instants =
-        ParameterUtil.rangeOfInstants(BACKFILL_2.start(), BACKFILL_2.end(), BACKFILL_2.schedule());
-    instants.forEach(
-        instant ->
-            verify(triggerListener).event(workflow, Trigger.backfill(BACKFILL_2.id()), instant));
-
-    final Backfill completedBackfill =
-        BACKFILL_2.builder().nextTrigger(BACKFILL_2.end()).allTriggered(true).build();
-    verify(storage).storeBackfill(completedBackfill);
-  }
-
-  @Test
-  public void shouldNotTriggerCompletedBackfillsAndUpdateCompleted() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    Backfill backfillWithNoPartitionsLeft = BACKFILL_2.builder().nextTrigger(BACKFILL_2.end()).build();
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(backfillWithNoPartitionsLeft));
-
-    scheduler.tick();
-
-    verifyZeroInteractions(triggerListener);
-    final Backfill completedBackfill = backfillWithNoPartitionsLeft.builder().allTriggered(true).build();
-    verify(storage).storeBackfill(completedBackfill);
-  }
-
-  @Test
-  public void shouldNotTriggerBackfillsIfResourceLimit() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(BACKFILL_1));
-
-    stateManager.initialize(
-        RunState.fresh(
-            WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T22")));
-    stateManager.receive(
-        Event.triggerExecution(
-            WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T22"),
-            Trigger.backfill("backfill-1")));
-    stateManager.initialize
-        (RunState.fresh(
-            WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T23")));
-    stateManager.receive(
-        Event.triggerExecution(
-            WorkflowInstance.create(WORKFLOW_ID1, "2016-12-02T23"),
-            Trigger.backfill("backfill-1")));
-
-    scheduler.tick();
-
-    verifyZeroInteractions(triggerListener);
-  }
-
-  @Test
-  public void shouldNotTriggerBackfillsWithMissingWorkflows() throws Exception {
-    setUp(5);
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(BACKFILL_1));
-
-    scheduler.tick();
-
-    verifyZeroInteractions(triggerListener);
-  }
-
-  @Test
-  public void shouldTriggerAndUpdateBackfillsSynchronously() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(BACKFILL_1));
-
-    // Collect unfinished triggering futures as the trigger listener is called
-    final BlockingQueue<CompletableFuture<Void>> triggerProcessingFutures =
-        new LinkedBlockingQueue<>();
-
-    when(triggerListener.event(
-        any(Workflow.class), any(Trigger.class), any(Instant.class)))
-        .then(a -> {
-          CompletableFuture<Void> processed = new CompletableFuture<>();
-          triggerProcessingFutures.add(processed);
-          return processed;
-        });
-
-    // Run a single tick of the scheduler
-    executor.execute(scheduler::tick);
-
-    final List<Instant> instants =
-        ParameterUtil.rangeOfInstants(BACKFILL_1.start(), BACKFILL_1.end(),
-            workflow.configuration().schedule());
-
-    // Go through each expected trigger sequentially and verify that the next partition is not
-    // triggered before the future for the previous partition trigger is completed
-    for (int i = 0; i < BACKFILL_1.concurrency(); i++) {
-      final Instant instant = instants.get(i);
-
-      final CompletableFuture<Void> triggerProcessed = triggerProcessingFutures.poll(1, TimeUnit.MINUTES);
-      assertThat(triggerProcessingFutures.isEmpty(), is(true));
-
-      verify(triggerListener, timeout(60_000))
-          .event(any(Workflow.class), any(Trigger.class), eq(instant));
-
-      triggerProcessed.complete(null);
-      verify(storage, timeout(60_000))
-          .storeBackfill(BACKFILL_1.builder().nextTrigger(instants.get(i + 1)).build());
-    }
-  }
-
-  @Test
-  public void shouldNotContinueTriggeringBackfillsIfTriggerFails() throws Exception {
-    setUp(5);
-    final Workflow workflow = workflowUsingResources(WORKFLOW_ID1);
-    initWorkflow(workflow);
-    when(storage.backfills(anyBoolean())).thenReturn(Collections.singletonList(BACKFILL_1));
-
-    RuntimeException rootCause = new RuntimeException("trigger listener failure!");
-
-    when(triggerListener.event(
-        any(Workflow.class), any(Trigger.class), any(Instant.class)))
-        .thenThrow(rootCause);
-
-    try {
-      scheduler.tick();
-      fail();
-    } catch (Exception e) {
-      assertThat(Throwables.getRootCause(e), is(sameInstance(rootCause)));
-    }
-
-    verify(triggerListener, timeout(60_000))
-        .event(any(Workflow.class), any(Trigger.class), any(Instant.class));
-
-    verifyNoMoreInteractions(triggerListener);
   }
 
   @Test
